Added servo offset
The servo angles are now offset by a set amount, to allow for small
inaccuracies in mounting them.  The offsets are set to default values,
but the laptop has the ability to change the offsets by using command 3.
The uS range of the servos was also changed to match the measured
values, rather than the datasheet.